Move renderTexture pool from FilterSystem
